CBCopilot/src/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic>=2.11.5
email-validator>=2.0
bcrypt==4.2.1
python-multipart==0.0.20
httpx==0.28.1
aiosmtplib>=3.0
openpyxl>=3.1
# RAG (Sprint 5) — pin LlamaIndex inside a known-good range borrowed from HRDD.
# torch is installed CPU-only as a separate Dockerfile step BEFORE this file
# to keep CUDA/NVIDIA bloat (~5 GB) out of the image.
llama-index-core>=0.13,<0.15
llama-index-readers-file>=0.5
llama-index-embeddings-huggingface>=0.6
# BM25 retriever for hybrid search (Sprint 9 overhaul). 0.5+ tracks the
# llama-index-core 0.13 line that embeddings-huggingface 0.6+ also requires —
# the older 0.4 series pinned core to <0.13 which deadlocked pip resolution.
# Brings in `rank_bm25` as a pure-Python transitive dep.
llama-index-retrievers-bm25>=0.5
# ChromaDB embedded vector store (Sprint 10C). Single persistent collection at
# /app/data/chroma; HNSW index + native metadata filtering. Replaces
# LlamaIndex's SimpleVectorStore. `chromadb` brings its own pinned numpy /
# posthog / opentelemetry; `llama-index-vector-stores-chroma` is the LlamaIndex
# adapter wired into `rag_service`.
chromadb>=0.5,<2.0
llama-index-vector-stores-chroma>=0.4
sentence-transformers>=2.2
docx2txt>=0.8
watchdog>=4.0
# Sprint 16 — Structured Table Pipeline. pdfplumber extracts tables from vector
# PDFs (returns empty for scanned / image-only PDFs, which we accept as
# low-fidelity). Pure Python except pdfminer.six, no native extensions → no
# Dockerfile pre-download step needed; pip install is enough.
pdfplumber>=0.11
